chore(frontend/hooks): 🎨 useTheme in hooks folder
