+++ app/frontend/src/components/mode-toggle.tsx
@@ -5,7 +5,7 @@ import {
 	DropdownMenuItem,
 	DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
-import { useTheme } from "@/context/theme-provider";
+import { useTheme } from "@/hooks/use-theme";
 import { Moon, Sun } from "lucide-react";
 
 export function ModeToggle() {
